Abort a commit() if it was deemed failure by timeout.
Seems like the eval process shouldn't necessarily be interrupted.  Since the nature of the script being eval'd is unknown trying to kill it might not yield good things or perhaps it won't be killed reliably.  Of course, this means that if enough long-run scripts are eval'd then it will bring gremlin-server to its knees.