Switching to AsciiDoctor's RTD theme
Disabled DocBook output and DocBook-related stylesheets.  AsciiDoctor
now renders directly to HTML5.  I've configured
asciidoctor-maven-plugin to use the ReadTheDocs theme that ships with
the AsciiDoctor Stylesheet Factory
(https://github.com/asciidoctor/asciidoctor-stylesheet-factory).  The
TOC is on the right-hand side, and code highlighting is enabled.

These changes are all in accordance with Marko's wishes, except for
the specific choice of CodeRay for highlighting, which didn't come up.
Discussing this earlier with Marko on IM, turn-key CSS styles with
minimal need for tweaking, richer/fancier code highlighting than
XSLTHL's default, and a right-hand fixed TOC emerged as priorities for
TP3 docs; the customizability and platform-independence of DocBook
were not priorities.  Hence this commit.