Corrected bug in StrategyWrappedElement and MapStep.
Unsure if the problem in MapStep was a bug or not, but all tests pass given the change.  Fixing the problem in StrategyWrappedElement, where the results from the call to id() was still doing a toString() of the result made it so that all the ReadOnlyGraphStrategy tests started failing which led me to the MapStep problem.  By not evaluating the class in MapStep for equality, the branching logic ended up being incorrect and a change in class from Element to StrategyWrappedElement did not register.  As a result Strategy tests no longer worked.  The toString() that was there previously made it look like things were working because the id() was being toString'd which did not equal a Long in that equality check in MapStep.