Replace deprecated assembly:aggregates w/ :single
Maven-assembly-plugin's "aggregate" goal has been deprecated for a
while, and it's generally badly behaved.  It forks the build
repeatedly when I try to build the entire TP3 reactor project, and
this gets worse when it interacts with javadoc:aggregate, which also
badly behaved and repeatedly forks the project.

maven-assembly-plugin is easier to fix than javadoc, because they've
deprecated the :aggregate goal and provided the new :single goal which
is essentially functionally identical but doesn't fork.