Initial commit of javadoc-to-S3 pom config
This runs javadoc:aggregate and uploads the results to the same S3
bucket as the asciidocs when the profile "javadoc-upload" is enabled
by setting either -Djavadoc-upload or -Pjavadoc-upload when invoking
mvn.

This has at least one known annoyonce: javadoc:aggregate forks the
build, and source:jar also forks the build, so the javadocs get
generated twice.