AsciiDoc S3 upload config should not be inherited
This prevent TP3 submodules from trying to upload AsciiDocs to S3.
Only the parent module is responsible for that.  Having submodules
attempt to do it is harmless in that it generate an error or result in
multiple uploads -- the submodules think there are no files to be
uploaded and hence just do nothing -- it could be confusing.