Removed the static do-nothing open() method from Graph interface.
The original intent was to remind implementers that they needed to implement that method, but it caused conflicts in the Gremlin Console (the console would call the base class do-nothing implementation, which made no sense as the base should have been hidden).  Anyway, tests do a better job of reminding implementers of their duties, as not including that static method in a Graph implementation will not allow the tests to run without that.