Corrected problem in console where eval'ing the addition of an item to a list caused an error.
Not sure why this was happening.  Could be related to something different between how things were handled in TP2 with ToStringPipe versus how it is being handled now.  A clone() of the result seemed to fix the problem and for just iterating results toString, I can't think of why that would hurt too much.