Modified the manner in which OptOut annotations are validated to be more robust.
No longer limited to just structure tests.  No longer implemented as a unit test that can't be opted out of.  It is built directly into the validation of the the implemented test suite making it easier to centrally enforce.  Provide better error messages to implementers so that it's clear about what needs to be fixed.