Corrected neo4j plugin loading problems in the console #230
Used a manifest entry in the neo4j-gremlin jar to help control where individual jar dependencies get written for the console when the :install command is used.  This enabled the neo4j-lucene-index jar to be moved up a directory which enables neo4j to work properly when installed this way.