Fixed broken integration tests.
Interestingly this amounted to renaming the sum() method in sample.groovy to addItUp().  Not sure why that is, because sum() has been fine for the last five months.  Not clear why this is suddenly broken.  I can only think that somehow groovy is hanging on to some other reference to sum() in a cache somewhere, but that seems unlikely too given the tests we have to prevent such things....mysterious.