Append plugin dirs to $CLASSPATH alphabetically
This change makes gremlin.sh append the ext/<plugin-dir> elements to
the Console's $CLASSPATH in alphabetical order according to the
plugin-dir names.  The previous behavior was dependent on the
filesystem's underlying order of directory entries, which was not
generally meaningful or obvious.

This commit also changes gremlin.sh to use wildcard classpath elements
for plugins, e.g. "ext/hadoop-gremlin/*".  This assumes that
ext/<plugin-dir> is always flat, that is that it has no internal
subdirectories and all of the jars are immediate children of
<plugin-dir>, and that the runtime JVM is at least version 6.  The
latter is definitely true since TP3 requires Java 8, and I think the
former is also true given how gremlin-console installs plugins.

The benefit of switching from enumerating every jar in the classpath
to using wildcards is that it makes the classpath easier to comprehend
when debugging and helps ensure that we never run afoul of process
environment size limits on older Linux kernels.