Attempt to remove random failures in Generator tests again #162
Unsure if this is fixed as the failure is random, but this change basically hardcodes a seed value for the generator that should be guaranteed to generate two different graphs.  That's the point of the failure and the point of the test (that two graphs will be different given different seeds).