Refactor IO for better consistency and implementer customization.
Made a number of changes here to make kryo and graphson work in the same fashion.  Added a Mapper interface that wraps around the GraphSON/Kryo to create the low-level serializers.  Modified the builder interfaces on GraphSONReader/Writer so that mapper configuration was decoupled from the reader/writers themselves.  The Io interface now includes a creator for GraphSONObjectMapper and renamed some methods there for consistency.