Remove requirement for Variables and VertexProperty.Iterators to be registered #468
Need to think this one through some more.  If a graph doesn't implement those interfaces, then they don't need to be registered, but that's a good way to cheat the system.  Of course, it wouldn't be impossible to cheat this system anyway with the way things are currently set up - just pass "empty" implementations....