Replaced all* remaining [source,groovy] code blocks with [gremlin] code blocks.
* where applicable; code blocks, that make use of any plugin, contain manually modified results and
code blocks that are supposed to produce an error, were not modified.

Moreover the preprocessor scripts were optimized, so that they now produce an exit code != 0 in case of an
error. This should allow us to cancel the build process and show a message, that one or more samples in the
docs are no longer working.