Make path to plugins.txt ignore working dir
Mediator's path to plugins.txt includes
System.getProperty("user.dir"), and it seems to assume that it's
always executed out of the parent of the ext directory.  Running the
shell anywhere besides that parent results in plugins not getting
loaded since the Mediator can't find plugins.txt.

This commit modifies the gremlin.sh to pass the path to the ext
directory as a system property ("tinkerpop.ext").  Mediator now
prefers tinkerpop.ext when present, but falls back to user.dir when
it's not.