Moved result iteration into the gremlin executor thread in Gremlin Server.
In making this move iteration occurs off of the netty worker thread.  I think that should have been seen as a blocking operation all this time and by making this change that issue is effectively removed.  The IteratorHandler was kept in the netty pipeline in case some ops wanted to make use of it.  In those cases, downstream handlers relying on it should not try to pass large resultsets through lest it cause problems.  We may yet want to simply kill the class - will give that some more thought.