Big refactor on the process tests.
Added UseEngine annotation to mark a test as one using a particular TraversalEngine.Type.  This allowed the removal of the boolean that used to be used to validate whether a graph would run a test given its ability to support GraphComputer.  The process computer/standard suite enforces that annotation and yields error messages that make sure the tests are setup properly.