Fixed SerializationTest for after api refactoring.
Replaced most references to Traversal in favor of usage of only low-level API references.  Had to leave a couple of serialization tests with Traversal still in them because the tests were looking at Traversal oriented objects: Path and Metrics.  Perhaps we need a set of serialization tests for gremlin-process related objects.