Added event listener methods on Transaction.
This enables users to detect when a transaction was successfully committed or rolled back.  Note that the default implementation in (and general expectation) is that event notifications are bound to the current thread in which they are registered.  in other words there is no global notification of a commit/rollback.