Introduce the gremlin-shaded module that houses kryo.
Shading was introduced to gremlin-core for kryo to enable it to work with spark.  Unfortunately, it created problems higher in the stack where certain modules like gremlin-driver needed access to those kryo files directly.  While this seemed to work ok for Maven, intellij had many problems processing the shaded jars preventing test execution and basic compilation.  Introducing gremlin-shaded provides a static jar that gremlin-core can depend on for intellij's purposes and as long as the gremlin-shaded is "removed" from intellij, things seem to work pretty well.