Change the serialization model for custom identifiers given TINKERPOP3-581
Using a Map to represent an custom element in GraphSON doesn't work so well for non-JVM languages and doesn't enable the default serialization pattern for GraphSON to be set with embedTypes as false.  This commit sets up for greater support for non-JVM language bindings which must conform to the JSON type system.