Add tests to enforce proper filtering of vertices during iteration TINKERPOP3-581
Updated TinkerGraph.vertices() implementation a bit.  Pushed some of the checking performed in GraphStep for non-uniform id arguments into the Graph.vertices() implementation so that it is handled at that level.  Still need to clean up in GraphStep a bit as that validation code probably doesn't need to be there anymore.