Refactoring test suites a bit for better clarity.
Removed redundant and unclear constructor overloads with default. Using more explicit construction now.  Put test enforcement back in - that was lost at some point.  Changed the nature of the gremlinFlavorSuite flag to print a message to System.err as a warning instead of doing nothing and just letting the test execute.  Rexamined all the settings for that flag in each suite.