Add revised Io interface with IoRegistry.
Allows implementers to register all their serialization classes with the IoRegistry which the Io implementations can consult for custom serializers.  This should get rid of the old Graph.Io interface which had tighter coupling to the Reader/Writer implementations.  At this point the two Io interfaces are still in there side-by-side.