Add GraphReader.readObject and GraphWriter.writeObject.
Not all IO implementations can support this, but for those that can like Gryo it provides a nice abstraction for folks to swap in and out of implementations.  Of course, the preferred way is probably to use the Io.mapper() to get the raw serializer implementation when possible, but for lightweight uses that is probably not necessary.