Provides a fix for TINKERPOP3-660 with the GraphMLReader.
GraphMLReader was not able to properly read <edge> elements before <node> elements.  The <edge> would create/cache the vertex, but then when the <node> came along it would only find it in the cache and not set the properties.  Note that this comes with a limitation that the vertex label will not be set if the <edge> comes before the <vertex> as per the TinkerPop contract that vertex labels can not be changed.  That limitation has been documented.