Prevent the default jackson serialization from taking over when types aren't embedded.
If a graph has a custom serializer and types are not embedded that serializer should still be used if present as it likely has a better and more compact representation than what the standard jackson serializer will do.  Furthermore the default serializer may try to serialize fields that throw exceptions thus halting the serialization process.