Add "rebindings" argument to sessionless requests.
This is an optional argument that allows a request to rebind a global variable on the server to a different name.  In this way, it is possible to always write traversals with "g" on the client side as you simply need to supply a rebinding of "g" to a different global var on each request.