Revised evaluation of ids with has() TINKERPOP3-721
Fixed problem where toString() of id() in has() was good for equality but not so good when you did g.V().has(id,lt(10)).  Modified the approach so that the conversion via toString() only occurs if the user calls has() with String values.  It is otherwise treated as the value that was passed in.