Enable transaction test to Gremlin Server with neo4j
Neo4j is packaged in a similar fashion as is done with neo4j-gremlin.  It only depends on the "api" in test scope thus allowing compilation and standard testing to work. Transactional tests "assume" neo4j presenece and are ignored otherwise.  To run transactional tests over neo4j they must be run with the -DincludeNeo4j flag. Note that it is not possible to "deploy" with that flag set (similar to how neo4j-gremlin works).