Move the Credentials DSL to gremlin-groovy.
In this way, the DSL does not need to be "installed" in the console.  If it had stayed on the server the user would have had to ":install gremlin-server" which is kinda weird (especially for just one little class).  Perhaps this is the way to go for future DSLs that we might want to support - kinda fits here in a way. Plus there is the added bonus that the DSL undergoes testing in the GroovyEnvironmentSuite which means it gets validated against all the graphs out there that run that suite.