The "secure" config now include CompileStaticCompilerCustomizerProvider
This meant that there had to be changes to the init script itself which wasn't type safe.  It also meant providing a new way to promote bindings as unresolved variables are not allowed under CompileStatic and TypeChecked.  Kept the old method of doing promotion but added the ability for the init script to pass back a Map of bindings to become global bindings.  That seems to work well.  Will deprecate the old method in 3.1.x.