Removed bindings promotion from GremlinExecutor.
The revised approach that simply takes values from a Map returned from the init script simplifies the initialization of global bindings and introduces new flexibility over the old more limited method of binding promotion.  As a result, binding promotion no longer needs to be supported as it really should not be the recommended way to deal with this issue.