Improve support around cardinality on calls to GraphTraversal.property()
This change was done in relation to PartitionStrategy.  The big change is that if the cardinality is specified to property() then the property will not be folded into an AddVertex/Start/EdgeStep.  This lets the cardinality continue to be respected.  The alternative to this approach would be to push Cardinality down to the Parameters class, but that was much more involved.  The cost here is a possible optimization that a graph might be able to take advantage of with all mutations held under the AddVertex/Start/EdgeStep.  Added tests to better cover cardinality calls and tests that cover VertexProperty related tests on PartitionStrategy.