Allows SparkGraphComputer to Preserve Spark Contexts between Jobs
Previously every traversal on a SparkGraphComputer would require
creating a brand new Spark Context. This adds a lot of overhead for
doing a bunch of operations in a row. To alliveate this we allow the
option of passing in a spark.persistContext paramter as part of the
HadoopConf. When this is enabled Spark Contexts will not be closed at
the end of the job. The getOrCreate context then allows a new
SparkGraphComputer get a handle on a previously running context.