Dropped the path visualization feature.
This isn't a breaking change as the feature was never officially released.  I was giving the code a second look today and there were some things that were some bad inefficiencies in there that I don't think are easily fixed given the way gephi works. This feature probably shouldn't make it into the wild.