TINKERPOP3-910 Clears the session channel attribute between requests.
If this value (and any other channel attribute) isn't cleared it will hang around until the next request as they are global to the channel.  In this case, a first "in-session" request would succeed and put the channel in "session" mode.  A future sessionless request would arrive and execute its serialization process in the sessions threads as the channel still had the session state from the previous request.  This likely would have been easier to debug if the the onReadWrite stuff had been ThreadLocal as the in-session requests would have been in MANUAL mode and the sessionless request would have yielded a more clear error that the transaction was closed.  Instead, because it was AUTO the sessionless request auto-opened the tx.

The tests for this are in the link referenced above.  We'll need to merge those tests into this branch at some point to get this all to work together.