TINKERPOP3-885 Adjustments to work from @PommeVerte from his PR at #113
Removed two tests that didn't really test threaded transactions properly.  Removed the static modifier on close/readWrite member variables which would cause a problem if the same thread was working with multiple graph instances.  Removed setReadWrite and setClose from AbstractTransaction as they were redundnant to just onReadWrite and onClose (they offered no functionality in and of themselves). Altered the nature of AbstractThreadedTransaction a bit as threaded transactions are really "manual" by virtue of their creation.  The transaction is opened when you createThreadedTransaction() and really shouldn't be re-used after close.  An new one should be created.  We don't have those semantics enforced now, but that's typically how this feature has been used in the past.