TINKERPOP3-955 Revised the tests for serialization of Map entries.
The tests now more closely map to how Gremlin Server works where Map results are converted to entrySet().  This change also improves the consistency of GraphSON by getting rid of the "key"/"value" keys and just uses the value for the key of an entry as the key.  It also better complies with the default key serializer used in GraphSON.