Altered transaction management in Gremlin Server
Removed the approach that did a commit prior to iteration and serialization as this opened a new transaction and wasn't optimal.  Instead, introduced a Frame abstraction that allowed serialization to occur during iteration of results.  In this way, results could be serialized in the same thread (and thus the same transaction) and errors in some graphs could be avoided.  The commit now occurs after the last item is serialized in the result set.  The last message is only flushed to the client after the commit occurs which assures that the the client shouldn't be able to send it's next request prior to that one being fully done.