Used a independent loop control for result interation.
This solves a problem where a graph could automatically open a new transaction on the same request.  In the result iteration loop, a commit gets called if there are no results left, when the loop traverse back to the top of the while() to exit, it does a final check to Traversal.hasNext().  On some graphs, automatic transaction handling could open a new transaction there. subtle - oh so subtle.