Add some additional "unexpected" error handling in OpExecutorHandler.
The OpExecutorHandler was only handling the error expected to come out of OpProcessor execution.  Adding a general exception catch does a better job of notifying the client if something goes wrong.  While doing this I also improved the semantics of a test designed to ensure that the client doesn't just "hang" and do nothing in the event of server failure.