Make the GraphManager instance injectable into ServerGremlinExecutor
This refers to a discussion from the gremlin-users mailing list:
https://groups.google.com/forum/#!topic/gremlin-users/xYzCAQExvvM

It is currently not possible to use a standard way to inject a
fully configured graph. The way presented in this gist is only
a hack and should not be takes as the default way.
https://gist.github.com/xenji/e469abde2e0b80aadc4e

Adding a new constructor parameter is, from my point of view,
the most flexible and compatible change to enable the injection
without breaking old code.

I've run the whole test suite of the server before pushing without
failures.