General fix to allow server-side fails to be reconnected.
If a server is responding to the driver in an expected fashion, then the server is still valid, but it may have killed the channel.  If it killed the channel and the driver tries to reconnect on that channel then future requests on that connection would immediately fail and the host would then get marked as dead (when it really wasn't).  Made some basic adjustments here to allow a Connection facing this situation to replace itself in the pool with a fresh connection which would open a new channel on the server and thus allow the next request using it to actually work.