Hooked together a basic but function infrastructure for traversal op
The traversal op allows a serialized Traversal to be shipped to Gremlin Server which can then be deserialized for execution on the server.  That eliminates the need for the ScriptEngine compilation. This current commit is really sketchy and just demonstrates operations.  Needs a lot more work to be fully functional.