Stopped setting "first" to false on reset()
The clone() of a Traversal was forcing a call to reset() on the client side for RemoteGraph.  When the Traversal deserialized on the server AddVertexStep behaved as though it had already been iterated. It's not clear why this field was being set this way on reset() and all tests seem to pass, so we'll stick with removal of the setting of this field.