TINKERPOP-1239 Connections were not be returned to the connection pool properly.
Excessive failures quickly drained the pool and resulted in TimeoutException for the driver. The fix was fairly simple as some poorly placed curly braces (or lack of curly braces) didn't allow an important line of code to execute. Added a test to validate that things were behaving properly now. CTR