Added some test to enforce some existing transactional semantics.
Covers TINKERPOP-1059 and TINKERPOP-947. While these were labelled as "breaking", that's a bit of a stretch of the definition. Existing expected semantics are now being enforced by tests so it's possible that providers could see a failure in their test suite over this change, but it is highly unlikely given the semantics being asserted. CTR