TINKERPOP-1302 Added shutdown hook to close open remotes.
Console mediator close doesn't need a Future. Not sure what was going on here (maybe incomplete??), but there was no need to return a CompletableFuture from Mediator.close() as the close was synchronous. It's also probably not a good idea to leave remotes open as they can leak remote resources. Change to a straight synchronous approach. CTR