TINKERPOP-1242 Tried to harden ScriptEngineTest to prevent random fail
There seemed to be only one test that might have the potential for some sort of hanging scenario. Made some modifications there to improve reliability. Removed a call to Thread.yield() which I think went in there under a change to get the test "better" under windows - hoping the CountDownLatch will do a better job in keeping that reliable and working on Windows. CTR