TINKERPOP-1341 Add missing classes to GryoRegistrator
GryoRegistrator is missing registrations for three classes registered
in GryoSerializer:

* void
* Void
* CompactBuffer[]

This commit adds all three to GryoRegistrator.getExtraRegistrations().

So far, CompactBuffer[] is the only omission known to cause practical
problems, as detailed by Dylan Bethune-Waddell on the mailing list and
in the referenced issue.  It's not clear whether omitting Void and
void causes practical problems, but they're trivial, so the cost
associated with registering them is tiny.  Might as well.