Increased settings on Cluster creation in test.
The test uses a fair degree of parallelism and now that other bug fixes went into play the pool wasn't growing out of bounds as it was before in some cases, which meant that the Cluster in the test had to wait longer for connections which ultimately led to timeouts. By increasing the number of in-process requests and increasing pool sizes, the test seems to now pass more consistently.