TINKERPOP-1254 Support dropping traverser path information when it is no longer needed
This commit adds support for path retraction to increase the likelihood of bulking in OLTP and OLAP modes. Traversal analysis is performed during the application of the PrunePathStrategy to identify labels that may be dropped at various points in the traversal.  MatchStep also performs runtime analysis to determine which labels it can drop in addition to the labels identified during traversal strategy application.