Allowed "imports" to be assigned to GremlinScriptEngine.
This commit made some wide reaching changes and is basically the initial integration of GremlinScriptEngine implementations into the wider code base. Note that the notion of a ScriptEngine in TinkerPop now only refers to a "gremlin-enabled" ScriptEngine and not just any ScriptEngine. In other words, it is not possible to just load a vanilla Nashorn or Jython ScriptEngine implementation. GremlinScriptEngine instances are also no longer registered as ScriptEngine implementations to be loaded under the standard ScriptEngineManager. We could likely still allow that but I'm not seeing where it is useful to do so. As a result, references to pure "jython" or "nashorn" and usage of raw vanilla ScriptEngine instances have been removed.

Added a basic method for handling imports in a generic way. The Customizer interface is central to that. Basically the ScriptEngineManager provides a list of Customizer implementations to each GremlinScriptEngineFactory. When the factory creates a new GremlinScriptEngine it passes those Customizer implementations to it on construction. The GremlinScriptEngine implementation must then decide how to use those Customizers to do its "setup".  This is a fairly close match to the CompilerCustomizerProvider approach that already exists in gremlin-groovy, but Customizer is a bit more generalized to gremlin-core. Ultimately, we will want to deprecate the CompilerCustomizerProvider in favor of the gremlin-core Customizer.