DSP-1397 Fix StarGraph addEdge for self-edges
StarGraph has directionally-specific classes for its edges:

* StarInEdge
* StarOutEdge

StarIn(/Out)Edges belong only in a StarVertex's in(/out)Edges map.  We
see this assumption in StarVertex.applyGraphFilter, which invokes
`graphFilter.legalEdges(this)` and then invokes `instanceof
StarOutEdge` on each legal edge to determine its direction.

This assumption was violated by the StarVertex.addEdge method.  This
method special-cases self-edge addition.  This method first
unconditionally adds a StarOutEdge.  Then, it checks to see whether
its parameters indicate that it is adding a self-edge.  If this is so,
the method proceeds to add that same StarOutEdge instance to the
inEdges map.  This is the problem.  Adding a self-edge to a StarVertex
results in one StarOutEdge instance that belongs to both the inEdges
and outEdges map, which later causes the `applyGraphFilter` method's
`instanceof` checks to yield two out edges and no in edges on a
self-loop.