Removed support for automatic unrolling of traversers in driver.
This feature was largely internal and not documented for public consumption. It's primary use was only tied to the remoting traversals and didn't have much other useful function. This feature has been moved further away from the driver and into client side remote traversal processing where it makes more sense. The driver really shouldn't maintain a notion of what a Traverser is. It's job is to merely send requests and get responses.