Enabled bytecode to be represented as graphson string or Bytecode object in RequestMessage.
Ultimately, perhaps this should just be a GraphSON string as that will universally work across all languages. The problem with doing that right now is that without embedTypes(true) graphson won't work so well given that it is type lossy. I think we should just save that change for GraphSON 2.0.