Added a way to extend the "end" behavior of result iteration.
There seemed to be some kind of race condition where the side-effect cache would not get initialized in between the final message being sent back to the client and the request for the side-effect. Introducing this extension allows TraversalOpProcessor to add the traversal sideeffect to the cache just prior to flush of the last messge. CTR