Changed traversal interrupt test to have a single sideEffect
The dual .sideEffect().sideEffect() which was unecessary in the first place, started presenting problems after NoOpBarrierStrategy was modified recently. The NoOpBarrierStep was splitting the sideEffect steps with a NoOpBarrierStep which was preventing the state modification in the second sideEffect() to affect the first. This test should pass consistently now. CTR