TINKERPOP-1434 Block calls on "remote" traversal to get side-effects
This prevents calls for side-effects that are not yet ready on the server. This was generally a problem in tests, but usrs could have also run afoul of the issue if they decided to get low-level with the API. It is more likely that they would grab side-effects with a call to cap() so this would be a non-issue in that case.