TINKERPOP-1544 Return confirmation on session close
While this is an improvement, it really sorta fixes a bug that can prevent certain tests from passing. By returning a confirmation of session close, it ensures that certain tests will always pass without having to rely on timers to simulate the situation. This change should be non-breaking to driver providers as to this point they had to have been doing simple fire-and-forget. CTR