Cleaned up error processing on remote side-effects.
The old method for parsing error message text from the server was kinda ugly and after recent refactoring it is ultimately unecessary as the keys are all known prior to a call to get() going through. Calls to get() will trap bad keys and make the "remote" side-effects behave like local ones and simply throw the appropriate exception. CTR