GraphManager support opening of specific graphs
This changeset allows an implementor to open a specific graph. One may
use this concept to implement a dynamic graph cache.

Furthermore, to ensure that rebindings work as intended, i.e. the list
of graphs returned to the HttpGremlinEndpointHandler, or "open graphs",
must include the to-be-rebound-graph. This changeset includes a change
to return these rebound graphs specifically.

Similar story as above for the WebSockets class, StandardOpProcessor.

Similar story for sessions, SessionOpProcessor.

Furthermore, the serializers at server boot only need to be aware of the
graphs defined in the settings object, so the relevant change here is in
AbstractChannelizer.

Furthermore:

To encourage a GraphManager implementation whose modification of the
Map<String, Graph> object aligns more closely with accepted "Getter and
Setter" design patterns, we update the adding of graphs to the
GraphManager Map by calling the new `addGraph(String, Graph)` rather
than publicly editting it with a call to `getGraphs().put(String,
Graph)`.

Also added `addTraversalSource(String, TraversalSource) for same
reasons.

Also, updated BasicGraphManager according to the new specifications.