Allow some extra time to be sure the future completes
This test never fails on my system but on rare occassions it seems to fail for others. I can only guess it is a timing issue of sorts, but I'm not sure I see how to better test this. The point of the test is not related to timing, so adding this extra wait doesn't seem wrong. It's not as though the functionality is not being tested properly. We just need to validate that the future completes when all the items arrive to the ResultSet, so in that sense the addition of the wait also validates that. CTR