TINKERPOP-1565 Reverted GraphSON 3.0 changes
These changes were part of the original work on this branch to alter the structure of GraphSON 2.0. I'd migrated those changes to 3.0 to avoid breaking change. Since that time there has been a lot of work on and off this branch related to this work and it seems that including GraphSON 3.0 changes to format is just complicating this ticket. As such the ticket itself has been modified to just be about setting up for GraphSON 3.0 and at this point 3.0 is equivalent to 2.0.