Convert to wildcard imports for the console
The console, under the revised plugin system and CoreImports, took a massive performance hit by not using wildcard imports. Imports are appended to each line executed in the console and that seems to have some significant expense in execution. It wouldn't be so bad as it was (maybe), but the doc generation process was very expensive since every line of the docs is passed through the console when there is source code to be processed there and it slowed doc generation down to 7x the previous speed. I imagine other automation use cases around the console would have been affected as well. CTR